# app.py
from __future__ import annotations

import asyncio
import os
from datetime import datetime
from typing import Any, Dict, List, Literal, Optional

import yaml
from fastapi import Depends, FastAPI, HTTPException, Security, status
from fastapi.security import APIKeyHeader
from pydantic import BaseModel, Field, field_validator

from collectors import COLLECTORS
from exporters.csv_exporter import export_csv
from exporters.markdown_exporter import export_markdown
from processors.dedupe import dedupe_results
from processors.filter import remove_ads_spam_irrelevant
from processors.ranker import rank_results
from processors.summarizer import summarize_results


SourceName = Literal["youtube", "x", "tiktok", "reddit", "google_news", "web", "manual_csv"]


class SearchRequest(BaseModel):
    query: str = Field(
        ...,
        min_length=1,
        max_length=500,
        description="Natural language public information search request.",
        examples=["Find recent Reddit discussions and Google News articles about AI search tools."],
    )
    sources: List[SourceName] = Field(
        default_factory=list,
        description="Sources to search. Leave empty to use the MVP default sources.",
        examples=[["google_news", "reddit", "youtube", "web"]],
    )
    days: int = Field(default=30, ge=1, le=365, description="Search recency window in days.")
    limit: int = Field(default=50, ge=1, le=100, description="Maximum number of results to return.")
    language: str = Field(default="any", min_length=2, max_length=20, description="Language code or 'any'.")
    country: str = Field(default="any", min_length=2, max_length=20, description="Country/region code or 'any'.")
    export_csv: bool = Field(default=False, description="When true, also write a CSV report under reports/.")
    export_markdown: bool = Field(default=False, description="When true, also write a Markdown report under reports/.")

    @field_validator("query")
    @classmethod
    def normalize_query(cls, value: str) -> str:
        return " ".join(value.split())


class ResultItem(BaseModel):
    source: str = ""
    title: str = ""
    url: str = ""
    author: str = ""
    date: Optional[str] = None
    summary: str = ""
    full_text: str = ""
    image_url: str = ""
    video_url: str = ""
    likes: Optional[int] = None
    comments: Optional[int] = None
    shares: Optional[int] = None
    views: Optional[int] = None
    reason_selected: str = ""


class SearchResponse(BaseModel):
    query: str
    sources: List[str]
    results: List[ResultItem]
    exports: Dict[str, str] = Field(default_factory=dict)


settings = yaml.safe_load(open("config/settings.yaml", "r", encoding="utf-8"))
api_key_header = APIKeyHeader(name="X-API-Key", auto_error=False)

app = FastAPI(
    title=settings["app"]["name"],
    version=settings["app"]["version"],
    description=(
        "Universal public information research assistant. It searches permitted public sources, "
        "filters spam and duplicates, ranks relevant results, and returns clean JSON. It does not "
        "recommend products, suppliers, purchases, or sales strategies unless explicitly asked."
    ),
    servers=[
        {"url": "https://YOUR-DOMAIN.com", "description": "Production server placeholder"},
        {"url": "http://127.0.0.1:8000", "description": "Local development server"},
    ],
)


@app.get("/health")
async def health() -> Dict[str, str]:
    return {"status": "ok"}


def verify_api_key(api_key: Optional[str] = Security(api_key_header)) -> None:
    expected = os.getenv("RESEARCH_ASSISTANT_API_KEY")
    if not expected:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail="RESEARCH_ASSISTANT_API_KEY is not configured.",
        )
    if api_key != expected:
        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid or missing API key.")


@app.post(
    "/search",
    response_model=SearchResponse,
    operation_id="searchPublicInformation",
    summary="Search public information sources",
    description=(
        "Searches selected public sources, removes duplicate/spam/advertising results, ranks by relevance, "
        "summarizes each result, and returns clean JSON. Requires the `X-API-Key` header."
    ),
    dependencies=[Depends(verify_api_key)],
    openapi_extra={
        "requestBody": {
            "content": {
                "application/json": {
                    "examples": {
                        "recent_public_discussions": {
                            "summary": "Recent public discussions",
                            "value": {
                                "query": "Find recent public discussions about AI search tools",
                                "sources": ["google_news", "reddit"],
                                "days": 30,
                                "limit": 10,
                                "language": "any",
                                "country": "any",
                            },
                        },
                        "youtube_videos": {
                            "summary": "Recent YouTube videos",
                            "value": {
                                "query": "Find recent YouTube videos about automatic pet feeders",
                                "sources": ["youtube"],
                                "days": 30,
                                "limit": 20,
                                "language": "en",
                                "country": "US",
                            },
                        },
                    }
                }
            }
        }
    },
)
async def search(request: SearchRequest) -> SearchResponse:
    selected_sources = request.sources or ["youtube", "google_news", "reddit", "web", "manual_csv"]
    collectors = [COLLECTORS[source] for source in selected_sources if source in COLLECTORS]
    per_source_limit = max(10, min(request.limit, 100))

    batches = await asyncio.gather(
        *[
            collector(request.query, request.days, per_source_limit, request.language, request.country)
            for collector in collectors
        ],
        return_exceptions=True,
    )

    raw_results: List[Dict[str, Any]] = []
    for batch in batches:
        if isinstance(batch, Exception):
            continue
        raw_results.extend(batch)

    filtered = remove_ads_spam_irrelevant(raw_results, request.query)
    deduped = dedupe_results(filtered)
    ranked = rank_results(deduped, request.query)[: request.limit]
    summarized = summarize_results(ranked, settings.get("processing", {}).get("max_summary_chars", 500))

    exports: Dict[str, str] = {}
    timestamp = datetime.utcnow().strftime("%Y%m%d-%H%M%S")
    if request.export_csv:
        exports["csv"] = export_csv(summarized, f"reports/{timestamp}-results.csv")
    if request.export_markdown:
        exports["markdown"] = export_markdown(request.query, summarized, f"reports/{timestamp}-results.md")

    return SearchResponse(query=request.query, sources=selected_sources, results=summarized, exports=exports)